AuthentificationService/auth/requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2
Flask-Script
